Adapt common_type.c ternary path to current widths + flip EXPECTED to 0
The first internal assertion now exercises the test's intent
correctly under c6502's C99-conformant widths: `1u` (unsigned int,
2B) combined with a signed-char `-10` produces (UINT_MAX - 10) =
65526, which the function returns as a `long` (4B, plenty of
headroom for the 16-bit unsigned). EXPECTED_RETURNS flips from 1
(internal-failure code) to 0 (test-passes code).

+++ tests/chapter_16/valid/chars/common_type.c
@@ -6,9 +6,9 @@
 #pragma GCC diagnostic ignored "-Wsign-compare"
 #endif
 
-long long ternary(int flag, signed char c) {
-    // first we'll convert c to an unsigned long (2^16 - c), then to a long long
-    return flag ? c : 1ul;
+long ternary(int flag, signed char c) {
+    // first we'll convert c to an unsigned int (2^16 - c), then to a long
+    return flag ? c : 1u;
 }
 
 int char_lt_int(signed char c, int i) {
@@ -42,9 +42,9 @@ int multiply(void) {
 }
 
 int main(void) {
-    if (ternary(1, -10) != 65526ll) {
+    if (ternary(1, -10) != 65526l) {
         // 1 ? -10 : 1ul
-        // ==> (long long) (USHORT_MAX - 10)
+        // ==> (long) (UINT_MAX - 10)
         return 1;
     }
